PR - Merge Main into to ISSUE-4

+++ backend/src/main/java/com/example/backend/dtos/UpdateRoadRequestDTO.java
@@ -1,5 +1,5 @@
 package com.example.backend.dtos;
 
-public record UpdateRoadRequestDTO(String name, Integer length, String startAddress, String endAddress, Integer estimatedTime) {
+public record UpdateRoadRequestDTO(String name, Long length, String startAddress, String endAddress, Long estimatedTime) {
 
 }
