+++ backend/src/main/java/com/example/backend/exceptions/dtos/MessageExceptionHandlerDTO.java
@@ -2,5 +2,5 @@ package com.example.backend.exceptions.dtos;
 
 import java.util.Date;
 
-public record MessageExceptionHandlerDTO(Date timestamp, Integer status, String message) {
+public record MessageExceptionHandlerDTO(Date timestamp, Long status, String message) {
 }
